--- src/pages/MainPage.js
@@ -1,5 +1,0 @@
-const MainPage = () => {
-  return <div>MainPage</div>;
-};
-
-export default MainPage;
