+++ src/pages/PrivacyPage/PrivacyPage.js
