+++ src/styles/_layout.scss
@@ -10,6 +10,7 @@
   display: flex;
   flex-direction: column;
   min-height: 100vh;
+  background: #fcfcfc;
 }
 
 // header
