+++ src/styles/_variables.scss
@@ -1,4 +1,15 @@
+// colors
 $primary-color: #3692ff;
+$white: #fff;
+$gray50: #f9fafb;
+$gray100: #f3f4f6;
+$gray200: #e5e7eb;
+$gray400: #9ca3af;
+$gray500: #6b7280;
+$gray600: #4b5563;
+$gray700: #374151;
+$gray800: #1f2937;
+$gray900: #111827;
 
 $footer-bg: #111827;
